Fix FTBFS on ppc64el

+++ source/gx/terminix/terminal/terminal.d
@@ -1681,7 +1681,7 @@ private:
     /**
      * Handler ID for scroll-event
      */
-    size_t scrollEventHandlerId;
+    gulong scrollEventHandlerId;
 
     void initColors() {
         vteFG = new RGBA();
